+++ app/javascript/components/TextBanner.scss
