+++ app/javascript/components/TextBanner.tsx
