+++ app/javascript/modules/listings/GenericDirectory.tsx
