+++ app/javascript/packs/react_application.tsx
@@ -1,14 +1,14 @@
 import WebpackerReact from "webpacker-react"
 
-import "../pages/base.scss"
+import "../components/base.scss"
 import "core-js/stable"
 import "regenerator-runtime/runtime"
 
-import RentDirectory from "../pages/RentDirectory"
-import BuyDirectory from "../pages/BuyDirectory"
-import HomePage from "../pages/HomePage"
-import SignIn from "../pages/SignIn"
-import ListingDetail from "../pages/ListingDetail"
+import RentDirectory from "../pages/listings/for-rent"
+import BuyDirectory from "../pages/listings/for-sale"
+import HomePage from "../pages"
+import SignIn from "../pages/sign-in"
+import ListingDetail from "../pages/listings/listing-detail"
 import { getCurrentLanguage, loadTranslations } from "../util/languageUtil"
 
 const currentLanguage = getCurrentLanguage(window.location.pathname)
